terminar los checkboxes

+++ ProyectoTiendaPracticaExamen/app/models/client.php
@@ -3,7 +3,6 @@ include_once '../../config/ConexionTienda.php';
 
 class client
 {
-
     public function __construct(
         private int $clientId = 0,
         private string $clientName = '',
@@ -26,51 +25,65 @@ class client
         }
     }
 
-    public function insertClient(client $newUser): bool
+    static public function insertClient(client $insertClient): bool
     {
-        if (empty($newUser)) {
+        if (empty($updateClient)) {
             throw new InvalidArgumentException('Error insertando un cliente, verifique los datos.');
         }
         $conexion = openConexionTienda();
-        $result = $conexion->query("INSERT INTO clientes VALUES ('{$newUser->clientName}','{$newUser->clientSurname}');");
+        $result = $conexion->query("INSERT INTO clientes VALUES ('{$insertClient->clientName}','{$insertClient->clientSurname}');");
         closeConexionTienda($conexion);
         return $result;
     }
 
-    // static public function updateClient(client $updateClient, int $clientId): bool
+    static public function updateClient(client $updateClient): bool
+    {
+        if (empty($updateClient)) {
+            throw new InvalidArgumentException('Error insertando un cliente, verifique los datos.');
+        }
+        $conexion = openConexionTienda();
+        $result = $conexion->query("UPDATE clientes SET nombre='{$updateClient->clientName}', apellidos='{$updateClient->clientSurname}' WHERE clienteId={$updateClient->clientId};");
+        closeConexionTienda($conexion);
+        return $result;
+    }
+
+    // public function updateClient(): bool
     // {
-    //     if (empty($updateClient)) {
-    //         throw new InvalidArgumentException('Error insertando un cliente, verifique los datos.');
-    //     }
     //     $conexion = openConexionTienda();
-    //     $result = $conexion->query("UPDATE clientes SET nombre='{$updateClient->clientName}', apellidos='{$updateClient->clientSurname}' WHERE clienteId={$clientId};");
+    //     $result = $conexion->query("UPDATE clientes SET nombre='{$this->clientName}', apellidos='{$this->clientSurname}' WHERE clienteId={$this->clientId};");
     //     closeConexionTienda($conexion);
     //     return $result;
     // }
 
-    public function updateClient(): bool
-    {
-        $conexion = openConexionTienda();
-        $result = $conexion->query("UPDATE clientes SET nombre='{$this->clientName}', apellidos='{$this->clientSurname}' WHERE clienteId={$this->clientId};");
-        closeConexionTienda($conexion);
-        return $result;
-    }
-
-    public function deleteClient(int $clientId): bool
+    static public function deleteClient(int $clientId): bool
     {
         if ($clientId < 0) {
             throw new InvalidArgumentException('El campo id cliente no puede ser negativo.');
         }
-        $conexion = openConexionTienda();
-        // no se podrá eliminar clietes que previamente hayan hecho un pedido, es para cuidar la consistencia de los datos
-        $result = $conexion->query("DELETE FROM clientes WHERE clienteId={$clientId};"); 
-        closeConexionTienda($conexion);
-        return $result;
+        try {
+            $conexion = openConexionTienda();
+            //verificamos si el cliente ha hecho pedidos, y si es así no lo permitiremos eliminar 
+            $numPedidos = $conexion->query("SELECT COUNT(*) FROM pedidos WHERE clienteId={$clientId};");
+            if ($numPedidos->fetch_column() > 0) {
+                throw new Exception('No se puede eliminar un cliente que tiene pedidos asociados.');
+            }
+            // no se podrá eliminar clietes que previamente hayan hecho un pedido, es para cuidar la consistencia de los datos
+            $result = $conexion->query("DELETE FROM clientes WHERE clienteId={$clientId};");
+            return $result;
+        } catch (Exception $e) {
+            // Captura cualquier excepción y muestra un mensaje de error.
+            error_log("Error: " . $e->getMessage());
+            return false;
+        } finally {
+            //siempre cerraremos la conexion
+            closeConexionTienda($conexion);
+        }
     }
 
-    static public function selectClient(int $clientId) : ?array {
+    static public function selectClient(int $clientId): ?array
+    {
         if ($clientId < 0) {
-            throw new InvalidArgumentException('El campo Id cliente no puede estar vacío.');            
+            throw new InvalidArgumentException('El campo Id cliente no puede estar vacío.');
         }
         $conexion = openConexionTienda();
         $result = $conexion->query("SELECT * FROM clientes WHERE clienteId='{$clientId}';");
@@ -78,7 +91,7 @@ class client
             $data = $result->fetch_row(MYSQLI_ASSOC);
             closeConexionTienda($conexion);
             return $data;
-        }else{
+        } else {
             closeConexionTienda($conexion);
             return null;
         }
